Migrate GSM SignalStrength to WCDMA on HAL 1.0
For HAL versions 1.0 and 1.1, there is no option to report
WCDMA signal strength via the SignalStrength UNSOL. Thus,
when the device is on WCDMA, we migrate the GSM RSSI and BER
to the WCDMA structure to fix the SignalStrength API to
return the correct structures.

Also, remove a second initMocks from RILTest that was causing all
mocks to be reinitialized, breaking unit tests that relied on any
existing mocks set up in TelephonyTest.

Bug: 134547711
Test: RILTest#testFixupSignalStrength10
Change-Id: I2701106d030511f41f657df5b85957476db80822
Signed-off-by: Josh Fox (XlxFoXxlX) <joshfox87@gmail.com>